requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
datetime
pymongo[srv]==3.6
pymongo
datetime
scikit-learn
PyYAML
ruamel.yaml
# -e .